requirements.txt
pytorch>=1.4
torchvision>=0.5
scikit-learn>=0.21.2
tqdm>=4.32.1
numpy>=1.16.4
scipy>=1.3.1
h5py>=2.9.0